Remove array type constructor template parameters
Right now, we determine the array (i.e. vector) type constructor through
a higher-kinded template parameter to most of our objects, but this is
somewhat inelegant. Instead, we will need to consolidate this to a
single source of truth, but that's for a later stage. For now, we get
rid of all of the type constructors and fix them to be `std::array`.