Encode storage data using output vectors
Currently, storage types encode the data they store using separate
dimension counts and types, but this is not in keeping with the output
vector types we use elsewhere. Thus, we convert the storage types to use
output vectors like the rest of the code. This is part of a larger
campagin to change how storage types work.