Simplify field backend concept type aliases
Previously, backends were required - by concept - to declare the input
and output types of both their covariant and contravariant parts.
However, this does not necessarily make sense, as the output of the
contravariant part and the input types of the covariant types are
essentially hidden from the outside world. This commit removes these
requirements from the concept, and removes the type aliases from classes
that had them for no good reason.